+++ src/js/records/entities/asset.record.js
@@ -1,6 +1,8 @@
 import { ASSET_POLICIES } from '@tokend/js-sdk'
 import { documentsManager } from '@/api'
 import safeGet from 'lodash/get'
+import { ASSET_STATES } from '@/js/const/asset.const'
+import moment from 'moment'
 
 export class AssetRecord {
   constructor (record = {}, balances = []) {
@@ -49,6 +51,9 @@ export class AssetRecord {
     this.stellarDeposit = safeGet(record, 'details.stellar.deposit') || false
 
     this.description = safeGet(record, 'details.description') || ''
+    this.state = safeGet(record, 'state.value')
+
+    this.expirationDate = this._getExpirationDate()
   }
 
   _getBalance (balances) {
@@ -83,6 +88,15 @@ export class AssetRecord {
 
   _policy () {
     return this._policies().reduce((s, p) => s | p, 0)
+  }
+
+  _getExpirationDate () {
+    const timestamp = safeGet(this._record, 'details.expiresAt')
+    if (timestamp) {
+      return moment.unix(timestamp).toISOString()
+    } else {
+      return ''
+    }
   }
 
   get record () {
@@ -145,5 +159,9 @@ export class AssetRecord {
     } else {
       return ''
     }
+  }
+
+  get isActive () {
+    return this.state === ASSET_STATES.active
   }
 }