+++ src/vue/pages/Customers.vue
@@ -87,9 +87,9 @@ export default {
   }),
 
   computed: {
-    ...mapGetters({
-      ownedAssets: vuexTypes.ownedAssets,
-    }),
+    ...mapGetters([
+      vuexTypes.activeOwnedAssets,
+    ]),
   },
 
   watch: {
@@ -127,7 +127,7 @@ export default {
     },
 
     getSelectionOptions () {
-      return this.ownedAssets.map(asset => {
+      return this.activeOwnedAssets.map(asset => {
         return {
           name: asset.name,
           value: asset.code,